Refactor icon imports to use consistent directory structure
